Move exceptions inside commons
