+++ src/main/java/seedu/address/commons/exceptions/DataConversionException.java
@@ -1,4 +1,4 @@
-package seedu.address.exceptions;
+package seedu.address.commons.exceptions;
 
 /**
  * Represents an error during conversion of data from one format to another
