+++ src/main/java/seedu/address/commons/exceptions/IllegalValueException.java
@@ -1,4 +1,4 @@
-package seedu.address.exceptions;
+package seedu.address.commons.exceptions;
 
 /**
  * Signals that some given data does not fulfill some constraints.
