+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
 import seedu.address.commons.util.CollectionUtil;
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.person.*;
 import seedu.address.model.tag.*;
