+++ src/main/java/seedu/address/model/person/Address.java
@@ -2,7 +2,7 @@ package seedu.address.model.person;
 
 
 import seedu.address.commons.util.CollectionUtil;
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 
 /**
  * Represents a Person's address in the address book.
