+++ src/main/java/seedu/address/model/tag/Tag.java
@@ -2,7 +2,7 @@ package seedu.address.model.tag;
 
 
 import seedu.address.commons.util.CollectionUtil;
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 
 /**
  * Represents a Tag in the address book.
