+++ src/main/java/seedu/address/storage/AdaptedPerson.java
@@ -1,7 +1,7 @@
 package seedu.address.storage;
 
 import seedu.address.commons.util.CollectionUtil;
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 import seedu.address.model.person.*;
