+++ src/main/java/seedu/address/storage/StorageAddressBook.java
@@ -1,6 +1,6 @@
 package seedu.address.storage;
 
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 import seedu.address.model.ReadOnlyAddressBook;
