+++ src/main/java/seedu/address/storage/StorageManager.java
@@ -4,7 +4,7 @@ import com.google.common.eventbus.Subscribe;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.events.model.LocalModelChangedEvent;
 import seedu.address.events.storage.*;
-import seedu.address.exceptions.DataConversionException;
+import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.core.ComponentManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.ReadOnlyAddressBook;
