+++ src/main/java/seedu/address/storage/XmlFileStorage.java
@@ -1,7 +1,7 @@
 package seedu.address.storage;
 
 import seedu.address.commons.util.XmlUtil;
-import seedu.address.exceptions.DataConversionException;
+import seedu.address.commons.exceptions.DataConversionException;
 
 import javax.xml.bind.JAXBException;
 import java.io.File;
