+++ src/test/java/seedu/address/storage/StorageManagerTest.java
@@ -8,8 +8,8 @@ import org.junit.Test;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.events.model.LocalModelChangedEvent;
 import seedu.address.events.storage.SaveDataRequestEvent;
-import seedu.address.exceptions.DataConversionException;
-import seedu.address.exceptions.DuplicateTagException;
+import seedu.address.commons.exceptions.DataConversionException;
+import seedu.address.commons.exceptions.DuplicateTagException;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.AddressBook;
