+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -1,6 +1,6 @@
 package seedu.address.testutil;
 
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Person;
