+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -17,7 +17,7 @@ import seedu.address.TestApp;
 import seedu.address.commons.util.FileUtil;
 import seedu.address.commons.core.OsDetector;
 import seedu.address.commons.util.XmlUtil;
-import seedu.address.exceptions.IllegalValueException;
+import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 import seedu.address.model.AddressBook;
